Remove UX error for UI boost
We'd like to throw an error if the user has
indicated they are receiving the $300 unemployment
income boost, but their value for other income
is less than $300, which would indicate one of
those values is incorrect. I'm not sure how these
errors can get passed back to the UI, so I've
disabled throwing an error and just assume the
meant to select "No" on the UI boost question.